requirements.txt
pandas
statsmodels
openai==0.28
pdfplumber
python-dotenv